Move more location methods to LocationManager. Shore up SpecimenEventManager. Remove many methods from SpecimenManager. Move ExtendedSpecimenRequestView, SpecimenRequest, and request requirement classes.
