+++ study/api-src/org/labkey/api/specimen/SpecimenEventManager.java
@@ -1,18 +1,34 @@
 package org.labkey.api.specimen;
 
+import org.jetbrains.annotations.NotNull;
+import org.labkey.api.data.Container;
+import org.labkey.api.data.Filter;
+import org.labkey.api.data.SimpleFilter;
+import org.labkey.api.data.TableInfo;
+import org.labkey.api.data.TableSelector;
+import org.labkey.api.query.FieldKey;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
 
-// Methods extracted from SpecimenManager to assist with migration. TODO: remove these or combine with SpecimenManager
+// Methods extracted from SpecimenManager to assist with migration. TODO: remove these or recombine with SpecimenManager
 public class SpecimenEventManager
 {
-    private static boolean skipAsProcessingLocation(SpecimenEvent event)
+    private static final SpecimenEventManager INSTANCE = new SpecimenEventManager();
+
+    private SpecimenEventManager()
     {
-        boolean allNullDates = event.getLabReceiptDate() == null && event.getStorageDate() == null && event.getShipDate() == null;
-        //
-        return allNullDates && !safeComp(event.getLabId(), event.getOriginatingLocationId());
     }
 
-    public static SpecimenEvent getFirstEvent(List<SpecimenEvent> dateOrderedEvents)
+    public static SpecimenEventManager get()
+    {
+        return INSTANCE;
+    }
+
+    public SpecimenEvent getFirstEvent(List<SpecimenEvent> dateOrderedEvents)
     {
         if (!dateOrderedEvents.isEmpty())
         {
@@ -26,7 +42,23 @@ public class SpecimenEventManager
         return null;
     }
 
-    public static SpecimenEvent getLastEvent(List<SpecimenEvent> dateOrderedEvents)
+    private boolean skipAsProcessingLocation(SpecimenEvent event)
+    {
+        boolean allNullDates = event.getLabReceiptDate() == null && event.getStorageDate() == null && event.getShipDate() == null;
+        //
+        return allNullDates && !safeComp(event.getLabId(), event.getOriginatingLocationId());
+    }
+
+    private static boolean safeComp(Object a, Object b)
+    {
+        if (a == null && b == null)
+            return true;
+        if (a == null || b == null)
+            return false;
+        return a.equals(b);
+    }
+
+    public SpecimenEvent getLastEvent(List<SpecimenEvent> dateOrderedEvents)
     {
         if (dateOrderedEvents.isEmpty())
             return null;
@@ -45,12 +77,51 @@ public class SpecimenEventManager
         return firstEvent != null ? firstEvent.getProcessedByInitials() : null;
     }
 
-    private static boolean safeComp(Object a, Object b)
+    public List<SpecimenEvent> getSpecimenEvents(@NotNull Vial vial)
     {
-        if (a == null && b == null)
-            return true;
-        if (a == null || b == null)
-            return false;
-        return a.equals(b);
+        SimpleFilter filter = new SimpleFilter(FieldKey.fromParts("VialId"), vial.getRowId());
+        return getSpecimenEvents(vial.getContainer(), filter);
+    }
+
+    public List<SpecimenEvent> getSpecimenEvents(List<Vial> vials, boolean includeObsolete)
+    {
+        if (vials == null || vials.size() == 0)
+            return Collections.emptyList();
+        Collection<Long> vialIds = new HashSet<>();
+        Container container = null;
+        for (Vial vial : vials)
+        {
+            vialIds.add(vial.getRowId());
+            if (container == null)
+                container = vial.getContainer();
+            else if (!container.equals(vial.getContainer()))
+                throw new IllegalArgumentException("All specimens must be from the same container");
+        }
+        SimpleFilter filter = new SimpleFilter();
+        filter.addInClause(FieldKey.fromString("VialId"), vialIds);
+        if (!includeObsolete)
+            filter.addCondition(FieldKey.fromString("Obsolete"), false);
+        return getSpecimenEvents(container, filter);
+    }
+
+    private List<SpecimenEvent> getSpecimenEvents(final Container container, Filter filter)
+    {
+        final List<SpecimenEvent> specimenEvents = new ArrayList<>();
+        TableInfo tableInfo = SpecimenSchema.get().getTableInfoSpecimenEvent(container);
+
+        new TableSelector(tableInfo, filter, null).forEachMap(map -> specimenEvents.add(new SpecimenEvent(container, map)));
+
+        return specimenEvents;
+    }
+
+    public List<SpecimenEvent> getDateOrderedEventList(Vial vial)
+    {
+        List<SpecimenEvent> eventList = new ArrayList<>();
+        List<SpecimenEvent> events = getSpecimenEvents(vial);
+        if (events == null || events.isEmpty())
+            return eventList;
+        eventList.addAll(events);
+        eventList.sort(SpecimenEventDateComparator.get());
+        return eventList;
     }
 }
